+++ etils/epath/testing.py
@@ -68,10 +68,16 @@ class _MockBackend(_backend_cls):
   def glob(self, path: PathLike) -> list[str]:
     return self._get_fn('glob')(path)
 
-  def makedirs(self, path: PathLike, *, exist_ok: bool = False) -> None:
+  def makedirs(
+      self, path: PathLike, *, exist_ok: bool = False, mode: int = 0o777
+  ) -> None:
+    del mode
     return self._get_fn('makedirs')(path, exist_ok=exist_ok)
 
-  def mkdir(self, path: PathLike, *, exist_ok: bool = False) -> None:
+  def mkdir(
+      self, path: PathLike, *, exist_ok: bool = False, mode: int = 0o777
+  ) -> None:
+    del mode
     return self._get_fn('mkdir')(path, exist_ok=exist_ok)
 
   def rmtree(self, path: PathLike) -> None: