+++ CHANGELOG.md
@@ -23,6 +23,8 @@ Changelog follow https://keepachangelog.com/ format.
             `;p`,...)
         *   Fixed auto-display when the line contain UTF-8 character
     *   Fix a bug for `ecolab.highlight_html` to escape HTML string.
+*   `epath`:
+    *   `path.mkdir` now supports `mode=` (for `os.path` and `gfile` backend).
 *   `epy`:
     *   `epy.lazy_imports()` support adhoc imports (will re-create the original
         `ecolab.adhoc` context when resolved)